changes to the read more clickable element + javscript scripting adjustments
